requirements.txt
pandas~=1.4.1
paho-mqtt~=1.6.1